Agregando cambios de agregados

+++ Backend/database/migrations/2025_09_23_214725_create_cliente_categorias_preferidas_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->integer('total_compras')->default(0);
             $table->decimal('total_gastado', 15, 2)->default(0);
             $table->decimal('porcentaje_gasto', 5, 2)->default(0);
-            $table->tinyInteger('ranking')->default(0);
+            $table->integer('ranking')->default(0);
             
             $table->timestamps();
